+++ RELEASE_NOTES.md
@@ -50,6 +50,7 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
 - When switching profiles the "Switch Filter" instruction will rematch the filter selection based on the name or the index
 - In framing tab you can now paste in full text coordinates into the RA/Dec text fields and they will be parsed into the separate boxes - e.g. when pasting the following string into the textbox the coordinates will be fully populated: `RA: 05h43m05s.90 DEC: +52°10′58″.0`
 - When sending location to the telescope the elevation is now handled separately and a different error message is shown
+- DitherAfterExposures will only trigger when the next item is an exposure item of type LIGHT
 
 ## Commandline Options
 - Multiple command line options have been added to be able to adjust some of the startup parameters for the application